m2/requirements.txt
asn1crypto==0.19.0
cffi==1.9.1
cryptography==1.7
enum34==1.1.6
idna==2.1
ipaddress==1.0.17
netifaces==0.10.5
numpy==1.11.0
ocspbuilder==0.10.2
oscrypto==0.17.2
py-cpuinfo==0.2.3
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.17
PyCPUID==0.4
pycrypto==2.6.1
PyKCS11==1.3.3
pyOpenSSL==16.2.0
six==1.10.0
